+++ Changelog.md
@@ -1,5 +1,15 @@
 # Change Log
 
+## v1.8.0
+
+---
+Release Date: **06.07.2021**
+
+- Added functions split (and freeze) a worksheet horizontally and vertically into panes
+- Added a property to set the visibility of a workbook
+- Added a property to set the visibility of worksheets
+- Added two examples in the demo for the introduced split, freeze and visibility functionalities
+
 ## v1.7.0
 
 ---